+++ Banking/View/WindowContent.xaml
@@ -10,6 +10,6 @@
     <Grid>
         <CornUI:CButton Width="100" Height="50" CornerRadius="10" BorderThickness="2, 2, 2, 2"
                     ShadowDirection="270" IconSize="40" FontSize="29" VerticalAlignment="Bottom" HorizontalAlignment="Left"/>
-        <CornUI:CTextbox Width="100" Height="50"/>
+        <CornUI:CTextbox Width="100" Height="30" ShadowThickness="10"/>
     </Grid>
 </UserControl>
