+++ CornUI/Controls/Normal/CButton.xaml
@@ -11,7 +11,7 @@
                    d:DesignHeight="20" d:DesignWidth="50"
                    Background="{Binding ElementName=root, Path=CurrentBackColor}"
                    BorderBrush="{Binding ElementName=root, Path=CurrentBorderColor}"
-                   BorderThickness="1,1,1,1" Text="Hi" Icon="/CornUI;component/Asset/Icon.png" IconSize="20">
+                   BorderThickness="1,1,1,1" Icon="/CornUI;component/Asset/Icon.png" IconSize="20">
     <local:CornControl.Resources>
         <ResourceDictionary>
             <Style TargetType="Border">
@@ -33,7 +33,7 @@
                           Direction="{Binding ElementName=root, Path=ShadowDirection}"
                           Opacity="{Binding ElementName=root, Path=ShadowOpacity}"
                           Color="{Binding ElementName=root, Path=CurrentBorderColor, Converter={StaticResource BrushAndColorConvertor}}"
-                          ShadowDepth="1" />
+                          ShadowDepth="0" />
     </local:CornControl.Effect>
     <DockPanel>
         <local:CIcon x:Name="icon"
