+++ CornUI/Controls/Normal/CTextbox.xaml
@@ -6,14 +6,43 @@
                    xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
                    xmlns:local="clr-namespace:CornUI.Controls.Normal"
                    mc:Ignorable="d" 
-                   d:DesignHeight="50" d:DesignWidth="100"
-                   CornerRadius="10">
+                   d:DesignHeight="30" d:DesignWidth="100"
+                   CornerRadius="5"
+                   BorderThickness="1,1,1,1"
+                   BorderBrush="{Binding ElementName=root, Path=CurrentBorderColor}"
+                   Background="{Binding ElementName=root, Path=CurrentBackColor}"
+                   MouseDown="ControlMouseDown"
+                   Cursor="IBeam">
+    <local:CornControl.Resources>
+        <ResourceDictionary>
+            <Style TargetType="Border">
+                <Setter Property="CornerRadius" Value="{Binding ElementName=root, Path=CornerRadius}"/>
+            </Style>
+            <ResourceDictionary.MergedDictionaries>
+                <ResourceDictionary Source="../ControlResources.xaml"/>
+            </ResourceDictionary.MergedDictionaries>
+        </ResourceDictionary>
+    </local:CornControl.Resources>
+    <local:CornControl.Effect>
+        <DropShadowEffect BlurRadius="{Binding ElementName=root, Path=ShadowThickness}" 
+                          Direction="{Binding ElementName=root, Path=ShadowDirection}"
+                          Opacity="{Binding ElementName=root, Path=ShadowOpacity}"
+                          Color="{Binding ElementName=root, Path=CurrentBorderColor, Converter={StaticResource BrushAndColorConvertor}}"
+                          ShadowDepth="0" />
+    </local:CornControl.Effect>
     <Grid>
         <TextBox x:Name="textBox"
-                 BorderBrush="{Binding ElementName=root, Path=CurrentBorderColor}"
-                 Background="{Binding ElementName=root, Path=CurrentBackColor}"
                  CaretBrush="{Binding ElementName=textBox, Path=Foreground}"
-                 PreviewKeyDown="textBox_PreviewKeyDown">
+                 Foreground="{Binding ElementName=root, Path=Foreground}"
+                 FontSize="{Binding ElementName=root, Path=FontSize}"
+                 FontWeight="{Binding ElementName=root, Path=FontWeight}"
+                 FontFamily="{Binding ElementName=root, Path=FontFamily}"
+                 FontStretch="{Binding ElementName=root, Path=FontStretch}"
+                 FontStyle="{Binding ElementName=root, Path=FontStyle}"
+                 Text="{Binding ElementName=root, Path=Text, UpdateSourceTrigger=PropertyChanged}"
+                 BorderThickness="0"
+                 VerticalAlignment="Center" 
+                 Background="Transparent">
             <TextBox.Resources>
                 <Style TargetType="Border">
                     <Setter Property="CornerRadius" Value="{Binding ElementName=root, Path=CornerRadius}"/>
@@ -25,8 +54,12 @@
                    x:Name="textBlock"
                    Text="{Binding ElementName=root, Path=DefaultText}"
                    Foreground="{Binding ElementName=root, Path=DefaulTextBrush}"
-                   Margin="5"
-                   MouseDown="TextBlocKMouseDown">
-        </TextBlock>
+                   FontSize="{Binding ElementName=root, Path=FontSize}"
+                   FontWeight="{Binding ElementName=root, Path=FontWeight}"
+                   FontFamily="{Binding ElementName=root, Path=FontFamily}"
+                   FontStretch="{Binding ElementName=root, Path=FontStretch}"
+                   FontStyle="{Binding ElementName=root, Path=FontStyle}"
+                   Margin="3"
+                   MouseDown="ControlMouseDown"/>
     </Grid>
 </local:CornControl>
